requirements.txt
annotated-types==0.7.0
anyio==4.8.0
black==25.1.0
buildozer==1.5.0
click==8.1.8
distlib==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.8
flake8==7.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.2.3
nvidia-nccl-cu12==2.25.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
ptyprocess==0.7.0
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
swagger_ui_bundle==1.1.0
threadpoolctl==3.5.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.1
watchfiles==1.0.4
websockets==14.2
xgboost==2.1.4
